refactor(mcp): split canopy-mcp lib.rs into domain modules
Pure structural refactor per PR #330 review: move the 17 tools out of a
single ~2200-line lib.rs into servers/groups/versions/fleet/backups/restore/
incidents modules, each with its own rmcp tool_router combined via
CanopyMcp::new(). Shared cross-module helpers (JSON shaping, id parsing,
lookups) move to a new util module. Tool names, descriptions, and output
shapes are unchanged.
