+++ crates/canopy-mcp/src/incidents.rs
@@ -1,0 +1,532 @@
+//! `find_incidents` / `get_incident` / `find_issues` / `get_issue` tools.
+
+use commons_types::{Uuid, issue::Severity};
+use database::{
+	issues::{Event, Incident, Issue, IssueListFilters},
+	server_groups::ServerGroup,
+	servers::Server,
+	slack_outbox::SlackOutbox,
+};
+use jiff::Timestamp;
+use rmcp::{
+	handler::server::wrapper::Parameters,
+	model::{CallToolResult, ErrorData as McpError},
+	tool, tool_router,
+};
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+	CanopyMcp,
+	util::{
+		group_names, mcp_err, not_found, ok_json, parse_opt_uuid, parse_uuid, since_from_days,
+		unique,
+	},
+};
+
+#[derive(Debug, Deserialize, JsonSchema)]
+pub struct FindIncidentsArgs {
+	/// Look back this many days; returns incidents that were open at any point in
+	/// the window (still open, or closed within it). Default 7.
+	pub since_days: Option<u32>,
+	/// Restrict to one group's id.
+	pub group_id: Option<String>,
+	/// Filter by status: `open` (not yet closed), `resolved` (operator-resolved),
+	/// or `all` (default).
+	pub status: Option<String>,
+	/// Max incidents to return (default 100).
+	pub limit: Option<i64>,
+}
+
+#[derive(Debug, Deserialize, JsonSchema)]
+pub struct IncidentIdArgs {
+	/// The incident's id.
+	pub incident_id: String,
+}
+
+#[derive(Debug, Deserialize, JsonSchema)]
+pub struct FindIssuesArgs {
+	/// Only currently-active, unresolved issues. Default true.
+	pub active_only: Option<bool>,
+	/// Filter to these severities: `critical`, `error`, `warning`, `info`, `debug`.
+	pub severities: Option<Vec<String>>,
+	/// Restrict to issues whose server is in this group's id.
+	pub group_id: Option<String>,
+	/// Restrict to one server's id.
+	pub server_id: Option<String>,
+	/// Only issues last seen within this many days.
+	pub since_days: Option<u32>,
+	/// Max issues to return (default 100).
+	pub limit: Option<i64>,
+}
+
+#[derive(Debug, Deserialize, JsonSchema)]
+pub struct IssueIdArgs {
+	/// The issue's id.
+	pub issue_id: String,
+}
+
+#[derive(Serialize)]
+struct IncidentSummary {
+	id: Uuid,
+	group_id: Uuid,
+	group_name: Option<String>,
+	/// `open` (not closed), `resolved` (operator-resolved), or `closed`.
+	status: &'static str,
+	opened_at: Timestamp,
+	closed_at: Option<Timestamp>,
+	resolved_at: Option<Timestamp>,
+	resolved_by: Option<String>,
+	resolved_reason: Option<String>,
+	/// Whether the incident ever escalated (a critical issue joined).
+	escalated: bool,
+	/// Whether the incident actually surfaced to operators (its Slack open
+	/// notice was delivered): it outlived the group's grace window, or it
+	/// escalated. Incidents that flapped shut within the grace never published.
+	/// Prefer counting `published` incidents over raw rows.
+	published: bool,
+	/// How long the incident was (or has been) open, in seconds.
+	open_duration_secs: i64,
+	issue_count: i64,
+	/// Raw count of status events the incident accumulated. NOT a measure of
+	/// duration or severity — a high count can be a sub-minute flap.
+	event_count: i64,
+}
+
+#[derive(Serialize)]
+struct IncidentList {
+	count: usize,
+	/// How many of `count` actually surfaced to operators (see `published`).
+	published_count: usize,
+	since: Timestamp,
+	incidents: Vec<IncidentSummary>,
+}
+
+#[derive(Serialize)]
+struct IncidentIssueOut {
+	issue_id: Uuid,
+	severity: Severity,
+	source: String,
+	r#ref: String,
+	description: Option<String>,
+	message: String,
+	active: bool,
+	server_id: Option<Uuid>,
+	server_name: Option<String>,
+	first_seen: Timestamp,
+	last_seen: Timestamp,
+	joined_at: Timestamp,
+	/// None = still attached to the incident.
+	left_at: Option<Timestamp>,
+}
+
+#[derive(Serialize)]
+struct IncidentDetail {
+	id: Uuid,
+	group_id: Uuid,
+	group_name: Option<String>,
+	status: &'static str,
+	opened_at: Timestamp,
+	closed_at: Option<Timestamp>,
+	resolved_at: Option<Timestamp>,
+	resolved_by: Option<String>,
+	resolved_reason: Option<String>,
+	escalated_at: Option<Timestamp>,
+	/// Whether the incident surfaced to operators (Slack open delivered).
+	published: bool,
+	open_duration_secs: i64,
+	created_at: Timestamp,
+	updated_at: Timestamp,
+	issues: Vec<IncidentIssueOut>,
+}
+
+#[derive(Serialize)]
+struct IssueSummary {
+	id: Uuid,
+	server_id: Option<Uuid>,
+	server_name: Option<String>,
+	group_id: Option<Uuid>,
+	source: String,
+	r#ref: String,
+	severity: Severity,
+	description: Option<String>,
+	message: String,
+	active: bool,
+	first_seen: Timestamp,
+	last_seen: Timestamp,
+	resolved_at: Option<Timestamp>,
+	snoozed_until: Option<Timestamp>,
+}
+
+#[derive(Serialize)]
+struct IssueList {
+	count: usize,
+	issues: Vec<IssueSummary>,
+}
+
+#[derive(Serialize)]
+struct EventOut {
+	created_at: Timestamp,
+	occurred_at: Option<Timestamp>,
+	severity: Severity,
+	description: Option<String>,
+	message: String,
+	active: bool,
+	occurrences: i32,
+	last_seen: Timestamp,
+}
+
+#[derive(Serialize)]
+struct IncidentRefOut {
+	incident_id: Uuid,
+	opened_at: Timestamp,
+	closed_at: Option<Timestamp>,
+}
+
+#[derive(Serialize)]
+struct IssueDetail {
+	id: Uuid,
+	server_id: Option<Uuid>,
+	server_name: Option<String>,
+	group_id: Option<Uuid>,
+	source: String,
+	r#ref: String,
+	severity: Severity,
+	description: Option<String>,
+	message: String,
+	active: bool,
+	first_seen: Timestamp,
+	last_seen: Timestamp,
+	resolved_at: Option<Timestamp>,
+	resolved_by: Option<String>,
+	resolved_reason: Option<String>,
+	snoozed_until: Option<Timestamp>,
+	recent_events: Vec<EventOut>,
+	incidents: Vec<IncidentRefOut>,
+}
+
+#[tool_router(router = incidents_router, vis = "pub(crate)")]
+impl CanopyMcp {
+	#[tool(
+		description = "List incidents that were open at any point in a recent window (default last \
+		               7 days), optionally for one group. Use this for 'incidents open in the past \
+		               week'.\n\n\
+		               IMPORTANT for summaries/ranking: count `published` incidents, not raw rows. \
+		               The window includes a large volume of sub-grace flapping (health checks that \
+		               recover/refire, alerts that self-clear in under a minute) that was recorded \
+		               but never surfaced to anyone. An incident is `published` only if its Slack \
+		               open notice was delivered: it stayed open past the group's grace window \
+		               (slack_open_delay, ~3 min by default) OR it escalated (a critical issue \
+		               joined, which bypasses the grace). `event_count` is raw status-event churn \
+		               and does NOT track duration or severity — a high-event incident can be a \
+		               sub-minute flap. A high count dominated by unpublished short-lived rows \
+		               usually means a twitchy alert/health-check threshold, not a real outage. \
+		               `published_count` gives the surfaced subset directly."
+	)]
+	async fn find_incidents(
+		&self,
+		Parameters(args): Parameters<FindIncidentsArgs>,
+	) -> Result<CallToolResult, McpError> {
+		let mut conn = self.conn().await?;
+		let since = since_from_days(args.since_days.unwrap_or(7));
+		let group = parse_opt_uuid(&args.group_id, "group_id")?;
+		let limit = args.limit.unwrap_or(100);
+		let status = args.status.as_deref().unwrap_or("all");
+
+		let incidents: Vec<Incident> = Incident::list_open_since(&mut conn, since, group, limit)
+			.await
+			.map_err(mcp_err)?
+			.into_iter()
+			.filter(|i| match status {
+				"open" => i.closed_at.is_none(),
+				"resolved" => i.resolved_at.is_some(),
+				_ => true,
+			})
+			.collect();
+
+		let group_names = group_names(
+			&mut conn,
+			&unique(incidents.iter().map(|i| i.server_group_id)),
+		)
+		.await?;
+		let ids: Vec<Uuid> = incidents.iter().map(|i| i.id).collect();
+		let stats = Incident::stats_for(&self.db, &ids).await.map_err(mcp_err)?;
+		let published = SlackOutbox::delivered_open_ids(&mut conn, &ids)
+			.await
+			.map_err(mcp_err)?;
+
+		let summaries: Vec<IncidentSummary> = incidents
+			.iter()
+			.map(|i| {
+				let s = stats.get(&i.id);
+				IncidentSummary {
+					id: i.id,
+					group_id: i.server_group_id,
+					group_name: group_names.get(&i.server_group_id).cloned(),
+					status: incident_status(i),
+					opened_at: i.opened_at,
+					closed_at: i.closed_at,
+					resolved_at: i.resolved_at,
+					resolved_by: i.resolved_by.clone(),
+					resolved_reason: i.resolved_reason.clone(),
+					escalated: i.escalated_at.is_some(),
+					published: published.contains(&i.id),
+					open_duration_secs: open_duration_secs(i),
+					issue_count: s.map_or(0, |s| s.issue_count),
+					event_count: s.map_or(0, |s| s.event_count),
+				}
+			})
+			.collect();
+
+		ok_json(&IncidentList {
+			count: summaries.len(),
+			published_count: summaries.iter().filter(|s| s.published).count(),
+			since,
+			incidents: summaries,
+		})
+	}
+
+	#[tool(
+		description = "Full detail for one incident: timing, status, and the issues attached to it \
+		               (with their severities and messages)."
+	)]
+	async fn get_incident(
+		&self,
+		Parameters(args): Parameters<IncidentIdArgs>,
+	) -> Result<CallToolResult, McpError> {
+		let mut conn = self.conn().await?;
+		let id = parse_uuid(&args.incident_id, "incident_id")?;
+		let Ok((incident, rows)) = Incident::get_with_issues(&mut conn, id).await else {
+			return Ok(not_found(format!("no incident with id {id}")));
+		};
+		let group = ServerGroup::get_by_id(&mut conn, incident.server_group_id)
+			.await
+			.ok();
+		let published = SlackOutbox::delivered_open_ids(&mut conn, &[incident.id])
+			.await
+			.map_err(mcp_err)?
+			.contains(&incident.id);
+		let names = Server::names_by_ids(
+			&mut conn,
+			&unique(rows.iter().filter_map(|(_, i)| i.server_id)),
+		)
+		.await
+		.map_err(mcp_err)?;
+
+		let issues = rows
+			.iter()
+			.map(|(link, iss)| IncidentIssueOut {
+				issue_id: iss.id,
+				severity: iss.severity,
+				source: iss.source.clone(),
+				r#ref: iss.r#ref.clone(),
+				description: iss.description.clone(),
+				message: iss.message.clone(),
+				active: iss.active,
+				server_id: iss.server_id,
+				server_name: iss
+					.server_id
+					.and_then(|s| names.get(&s))
+					.and_then(|(n, _)| n.clone()),
+				first_seen: iss.first_seen,
+				last_seen: iss.last_seen,
+				joined_at: link.joined_at,
+				left_at: link.left_at,
+			})
+			.collect();
+
+		ok_json(&IncidentDetail {
+			id: incident.id,
+			group_id: incident.server_group_id,
+			group_name: group.as_ref().map(|g| g.name.clone()),
+			status: incident_status(&incident),
+			opened_at: incident.opened_at,
+			closed_at: incident.closed_at,
+			resolved_at: incident.resolved_at,
+			resolved_by: incident.resolved_by.clone(),
+			resolved_reason: incident.resolved_reason.clone(),
+			escalated_at: incident.escalated_at,
+			published,
+			open_duration_secs: open_duration_secs(&incident),
+			created_at: incident.created_at,
+			updated_at: incident.updated_at,
+			issues,
+		})
+	}
+
+	#[tool(
+		description = "List issues across the fleet, filtered by active state, severity, group, \
+		               server, and recency. Issues are the per-(server,source,ref) events that make \
+		               up incidents."
+	)]
+	async fn find_issues(
+		&self,
+		Parameters(args): Parameters<FindIssuesArgs>,
+	) -> Result<CallToolResult, McpError> {
+		let mut conn = self.conn().await?;
+		let severities = parse_severities(&args.severities)?;
+		let group = parse_opt_uuid(&args.group_id, "group_id")?;
+		let server = parse_opt_uuid(&args.server_id, "server_id")?;
+		let since = args.since_days.map(since_from_days);
+		let limit = args.limit.unwrap_or(100);
+
+		let mut issues = Issue::list(
+			&mut conn,
+			IssueListFilters {
+				active_only: args.active_only.unwrap_or(true),
+				severities,
+				server_group_id: group,
+				since,
+			},
+			limit,
+		)
+		.await
+		.map_err(mcp_err)?;
+		if let Some(sid) = server {
+			issues.retain(|i| i.server_id == Some(sid));
+		}
+
+		let names = Server::names_by_ids(
+			&mut conn,
+			&unique(issues.iter().filter_map(|i| i.server_id)),
+		)
+		.await
+		.map_err(mcp_err)?;
+		let summaries: Vec<IssueSummary> =
+			issues.iter().map(|i| issue_summary(i, &names)).collect();
+		ok_json(&IssueList {
+			count: summaries.len(),
+			issues: summaries,
+		})
+	}
+
+	#[tool(
+		description = "Full detail for one issue: its fields, recent events, and the incidents it \
+		               is or was part of."
+	)]
+	async fn get_issue(
+		&self,
+		Parameters(args): Parameters<IssueIdArgs>,
+	) -> Result<CallToolResult, McpError> {
+		let mut conn = self.conn().await?;
+		let id = parse_uuid(&args.issue_id, "issue_id")?;
+		let Ok(issue) = Issue::get_by_id(&mut conn, id).await else {
+			return Ok(not_found(format!("no issue with id {id}")));
+		};
+		let events = Event::list_for_issue(&mut conn, id, 0, 20)
+			.await
+			.map_err(mcp_err)?;
+		let inc = Incident::for_issues(&mut conn, &[id])
+			.await
+			.map_err(mcp_err)?;
+		let server_name = match issue.server_id {
+			Some(sid) => Server::names_by_ids(&mut conn, &[sid])
+				.await
+				.map_err(mcp_err)?
+				.get(&sid)
+				.and_then(|(n, _)| n.clone()),
+			None => None,
+		};
+
+		let recent_events = events
+			.iter()
+			.map(|e| EventOut {
+				created_at: e.created_at,
+				occurred_at: e.occurred_at,
+				severity: e.severity,
+				description: e.description.clone(),
+				message: e.message.clone(),
+				active: e.active,
+				occurrences: e.occurrences,
+				last_seen: e.last_seen,
+			})
+			.collect();
+		let incidents = inc
+			.get(&id)
+			.into_iter()
+			.flatten()
+			.map(|r| IncidentRefOut {
+				incident_id: r.incident_id,
+				opened_at: r.opened_at,
+				closed_at: r.closed_at,
+			})
+			.collect();
+
+		ok_json(&IssueDetail {
+			id: issue.id,
+			server_id: issue.server_id,
+			server_name,
+			group_id: issue.server_group_id,
+			source: issue.source.clone(),
+			r#ref: issue.r#ref.clone(),
+			severity: issue.severity,
+			description: issue.description.clone(),
+			message: issue.message.clone(),
+			active: issue.active,
+			first_seen: issue.first_seen,
+			last_seen: issue.last_seen,
+			resolved_at: issue.resolved_at,
+			resolved_by: issue.resolved_by.clone(),
+			resolved_reason: issue.resolved_reason.clone(),
+			snoozed_until: issue.snoozed_until,
+			recent_events,
+			incidents,
+		})
+	}
+}
+
+/// How long the incident was (or has been) open, in seconds.
+fn open_duration_secs(i: &Incident) -> i64 {
+	let end = i.closed_at.unwrap_or_else(Timestamp::now);
+	end.duration_since(i.opened_at).as_secs().max(0)
+}
+
+fn incident_status(i: &Incident) -> &'static str {
+	if i.resolved_at.is_some() {
+		"resolved"
+	} else if i.closed_at.is_some() {
+		"closed"
+	} else {
+		"open"
+	}
+}
+
+fn parse_severities(v: &Option<Vec<String>>) -> Result<Option<Vec<Severity>>, McpError> {
+	match v {
+		Some(list) if !list.is_empty() => {
+			let mut out = Vec::with_capacity(list.len());
+			for s in list {
+				out.push(s.parse::<Severity>().map_err(|_| {
+					McpError::invalid_params(format!("invalid severity: {s}"), None)
+				})?);
+			}
+			Ok(Some(out))
+		}
+		_ => Ok(None),
+	}
+}
+
+fn issue_summary(
+	i: &Issue,
+	names: &std::collections::HashMap<Uuid, (Option<String>, Option<String>)>,
+) -> IssueSummary {
+	IssueSummary {
+		id: i.id,
+		server_id: i.server_id,
+		server_name: i
+			.server_id
+			.and_then(|s| names.get(&s))
+			.and_then(|(n, _)| n.clone()),
+		group_id: i.server_group_id,
+		source: i.source.clone(),
+		r#ref: i.r#ref.clone(),
+		severity: i.severity,
+		description: i.description.clone(),
+		message: i.message.clone(),
+		active: i.active,
+		first_seen: i.first_seen,
+		last_seen: i.last_seen,
+		resolved_at: i.resolved_at,
+		snoozed_until: i.snoozed_until,
+	}
+}
